+++ src/stores/ui.store.ts
@@ -1,8 +1,6 @@
 import { createStore } from "solid-js/store";
 import { createSignal } from "solid-js";
 import { listen } from "@tauri-apps/api/event";
-
-export { showToast, dismissToast, toasts, type ToastKind, type Toast } from "@/components/ui/Toast";
 import { showToast } from "@/components/ui/Toast";
 
 export type MainTab = "build" | "logcat" | "layout";
